combine string util into utilities file
